requirements.txt
h5py==2.10.0
matplotlib==3.5.2
numpy==1.18.5
pandas==1.0.5
Pillow==9.0.0
pycocotools_windows==2.0.0.2
scikit_learn==0.23.1
torch==1.8.0+cu111
torchvision==0.9.0+cu111
